requirements.txt
Django==3.2.3
django-import-export==2.5.0
django-summernote==0.8.11.6
djangorestframework==3.12.4
Markdown==3.3.4
MarkupPy==1.14
Pillow==8.3.1
sorl-thumbnail==12.7.0
gunicorn==20.1.0
psycopg2==2.9.1
requests==2.26.0
boto3==1.18.26
django-storages==1.11.1
django-cors-headers==3.8.0
